pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chromemind"
version = "0.1.0"
description = "Agentic Browser Knowledge Curator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "notion-client==2.2.1",
    "pydantic>=2.7.1",
    "aiohttp>=3.9.5",
    "google-genai>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
]

[project.scripts]
chromemind = "chromemind.cli:cli"

[tool.setuptools.packages.find]
include = ["chromemind*", "skills*", "agents*"]
